LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/ready-states/autoplay-with-slow-text-tracks-expected.txt

CONSOLE MESSAGE: line 31: TypeError: null is not an object (evaluating 'track.track.cues.length')
CONSOLE MESSAGE: line 31: TypeError: null is not an object (evaluating 'track.track.cues.length')
CONSOLE MESSAGE: line 31: TypeError: null is not an object (evaluating 'track.track.cues.length')
CONSOLE MESSAGE: line 31: TypeError: null is not an object (evaluating 'track.track.cues.length')
CONSOLE MESSAGE: line 31: TypeError: null is not an object (evaluating 'track.track.cues.length')

Harness Error (FAIL), message = TypeError: null is not an object (evaluating 'track.track.cues.length')
